requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
passlib==1.7.4
argon2-cffi==23.1.0
python-jose==3.3.0
email-validator==2.2.0
python-multipart==0.0.21

pytest==8.3.4
httpx==0.28.1